É agora que o pwm vai dar!

+++ RPI-codes/4-InitialServer/src/i2cFunctions.cpp
@@ -104,7 +104,7 @@ void I2cFunctions::readLoop(Data* database) {
                 case 3:
                     printf("Arduino %d ", xfer.rxBuf[0]);
                     printf("\tOccupancy: %d\n\n", xfer.rxBuf[2]);
-                    database->setOccupancyAtDesk(xfer.rxBuf[2], xfer.rxBuf[0]);
+                    database->setOccupancyAtDesk(xfer.rxBuf[2], (int)xfer.rxBuf[0]);
                     break;
                 // lux lower bound
                 case 4:
@@ -115,7 +115,7 @@ void I2cFunctions::readLoop(Data* database) {
                     lux = this->mapfloat(LuxValue, 0, 65536, 0, MAX_LUX);
                     printf("\tLux Lower Bound: %f\n\n", lux);
                     // Add value to database
-                    database->setLuxLowerBoundAtDesk(lux, xfer.rxBuf[0]);
+                    database->setLuxLowerBoundAtDesk(lux, (int)xfer.rxBuf[0]);
                     break;
                 // background lux
                 case 5:
@@ -126,7 +126,7 @@ void I2cFunctions::readLoop(Data* database) {
                     lux = this->mapfloat(LuxValue, 0, 65536, 0, MAX_LUX);
                     printf("\tBackground Lux: %f\n\n", lux);
                     // Add value to database
-                    database->setLuxExternalAtDesk(lux, xfer.rxBuf[0]);
+                    database->setLuxExternalAtDesk(lux, (int)xfer.rxBuf[0]);
                     break;
                 // control reference pwm
                 case 6:
@@ -134,7 +134,7 @@ void I2cFunctions::readLoop(Data* database) {
                     pwm = this->mapfloat(xfer.rxBuf[2], 0, 255, 0, 5);
                     printf("\tReference pwm: %f\n\n", pwm);
                     // Add value to database
-                    database->setLuxControlReference(pwm, xfer.rxBuf[0]);
+                    database->setLuxControlReference(pwm, (int)xfer.rxBuf[0]);
                     break;
             }
